kind of cracked the support for floats. Kind of.
it's complicated because there are just a few data structures that are
used for everything. I.e. for both the AST and the number of arguments
passed to a function. So there is some casting needed ever now and then.

Still to do: in theory arrays can be indexed by numbers bigger than
ints.

Also, floats are not parsed from the source file yet.

+++ Rhash.java
@@ -28,7 +28,7 @@ public class Rhash {
         return this;
     }
     
-    public final int get(int key, int defValue) {
+    public final double get(int key, int defValue) {
         Rv entry = getEntry(key, null);
         return entry != null ? entry.num : defValue;
     }
